plotfix: PlotFromTable needs to operate on table, not plot, because could use basic plot instead of plotcore Editor.
