+++ stats/cluster/clust_test.go
@@ -7,10 +7,14 @@ package cluster
 import (
 	"testing"
 
+	"cogentcore.org/core/base/iox/imagex"
 	"cogentcore.org/core/base/tolassert"
+	"cogentcore.org/lab/plot"
+	_ "cogentcore.org/lab/plot/plots"
 	"cogentcore.org/lab/stats/metric"
 	"cogentcore.org/lab/table"
 	"cogentcore.org/lab/tensor"
+	"github.com/alecthomas/assert/v2"
 )
 
 var clustres = `
@@ -30,9 +34,7 @@ var clustres = `
 func TestClust(t *testing.T) {
 	dt := table.New()
 	err := dt.OpenCSV("testdata/faces.dat", tensor.Tab)
-	if err != nil {
-		t.Error(err)
-	}
+	assert.NoError(t, err)
 	in := dt.Column("Input")
 	out := metric.Matrix(metric.L2Norm, in)
 
@@ -52,4 +54,17 @@ func TestClust(t *testing.T) {
 	exdists := []float64{0, 9.181170119179619, 5.534356355667114, 4.859933137893677, 3.464101552963257, 0, 0, 3.464101552963257, 0, 0, 3.464101552963257, 0, 0, 5.111664593219757, 4.640135824680328, 4, 0, 0, 3.464101552963257, 0, 0, 3.605551242828369, 0, 0}
 
 	tolassert.EqualTolSlice(t, exdists, dists, 1.0e-7)
+}
+
+func TestTableCluster(t *testing.T) {
+	dt := table.New()
+	err := dt.OpenCSV("testdata/faces.dat", tensor.Tab)
+	assert.NoError(t, err)
+
+	pt := table.New()
+	PlotFromTable(pt, dt, metric.MetricL2Norm, Min, "Input", "Name")
+	plt, err := plot.NewTablePlot(pt)
+	assert.NoError(t, err)
+	fnm := "table_cluster.png"
+	imagex.Assert(t, plt.RenderImage(), fnm)
 }
