+++ stats/cluster/plot.go
@@ -6,26 +6,24 @@ package cluster
 
 import (
 	"cogentcore.org/lab/plot"
-	"cogentcore.org/lab/plotcore"
 	"cogentcore.org/lab/stats/metric"
 	"cogentcore.org/lab/table"
 	"cogentcore.org/lab/tensor"
 )
 
-// PlotFromTable creates a cluster plot in given [plotcore.Editor]
-// using data from given data table, in column dataColumn, and labels from labelColumn,
-// with given distance metric and cluster metric functions.
-func PlotFromTable(plt *plotcore.Editor, dt *table.Table, distMetric metric.Metrics, clustMetric Metrics, dataColumn, labelColumn string) {
+// PlotFromTable creates a cluster plot in given plot data table
+// using data from given data table, in column dataColumn,
+// and labels from labelColumn, with given distance metric
+// and cluster metric functions.
+func PlotFromTable(pt *table.Table, dt *table.Table, distMetric metric.Metrics, clustMetric Metrics, dataColumn, labelColumn string) {
 	dm := metric.Matrix(distMetric.Func(), dt.Column(dataColumn))
 	labels := dt.Column(labelColumn)
 	cnd := Cluster(clustMetric, dm, labels)
-	pdt := table.New()
-	Plot(pdt, cnd, dm, labels)
-	plotcore.SetBasicStylers(pdt)
-	plot.Styler(pdt.Columns.At("Label"), func(s *plot.Style) {
+	Plot(pt, cnd, dm, labels)
+	plot.SetBasicStylers(pt)
+	plot.Styler(pt.Columns.At("Label"), func(s *plot.Style) {
 		s.On = true
 	})
-	plt.SetTable(pdt)
 }
 
 // Plot sets the rows of given data table to trace out lines with labels that
